Clean up compiler warnings for 64-bit systems where __u64 is a long and not a long long like on 32-bit systems.  From Evan Felix, PNL.

+++ lustre/lib/simple.c
@@ -85,7 +85,7 @@ struct dentry *simple_mkdir(struct dentry *dir, char *name, int mode)
         ENTRY;
 
         ASSERT_KERNEL_CTXT("kernel doing mkdir outside kernel context\n");
-        CDEBUG(D_INODE, "creating directory %*s\n", strlen(name), name);
+        CDEBUG(D_INODE, "creating directory %*s\n", (int)strlen(name), name);
         down(&dir->d_inode->i_sem);
         dchild = lookup_one_len(name, dir, strlen(name));
         if (IS_ERR(dchild))
